SettingsNewPrivacyPanel renamed to SettingsPrivacyPanel (#1588).
