+++ assets/strings.en.json
@@ -800,30 +800,6 @@
     "panel.events_schedule.button.details.title": "Details",
     "panel.events_schedule.button.details.hint": "",
 
-    "panel.settings.privacy.label.title":"My Privacy",
-    "panel.settings.privacy.label.set_your_privacy_level":"Set your privacy level",
-    "panel.settings.privacy.label.set_your_privacy_level.hint": "Header 1",
-    "panel.settings.privacy.label.update_your_privacy_level":"Update your privacy level",
-    "panel.settings.privacy.label.update_your_privacy_level.hint": "Header 1",
-    "panel.settings.privacy.label.some_details_have_changed": "Some details have changed",
-    "panel.settings.privacy.label.some_details_have_changed.hint": "Header 2",
-    "panel.settings.privacy.label.category.hint": "Header ",
-    "panel.settings.privacy.label.description":"Choose your privacy level with the slider below. You can review what information you share and change your setting at any time.",
-    "panel.settings.privacy.label.bottom_description":"Regardless of the privacy level anonymous data will be collected and used to optimize performance",
-    "panel.settings.privacy.button.set_privacy.title":"Set my Privacy",
-    "panel.settings.privacy.button.set_privacy.hint":"Set my privacy",
-    "panel.settings.privacy.button.set_privacy.disabled.title":"Scroll to Review",
-    "panel.settings.privacy.button.set_privacy.disabled.hint":"Scroll to Review",
-    "panel.settings.privacy.button.set_privacy.slider.hint":"Privacy Level",
-    "panel.settings.privacy.button.set_privacy.slider.increase":"increased to",
-    "panel.settings.privacy.button.set_privacy.slider.decrease":"decreased to",
-    "panel.settings.privacy.dialog.update_privacy.title":"New Privacy Settings",
-    "panel.settings.privacy.dialog.update_privacy.label1":"Your new privacy setting",
-    "panel.settings.privacy.dialog.update_privacy.label2":"requires us to make the following chages:",
-    "panel.settings.privacy.dialog.update_privacy.confirm":"Are you sure?",
-    "panel.settings.privacy.dialog.update_privacy.yes":"Yes",
-    "panel.settings.privacy.dialog.update_privacy.no":"No",
-
     "panel.settings.manage_interests.title":"Manage My Interests",
     "panel.settings.manage_interests.instructions.tap":"Tap the",
     "panel.settings.manage_interests.instructions.follow": " to follow the tags that interest you most",
@@ -1227,34 +1203,34 @@
     "panel.settings.personal_information.label.delete_message.description3":"delete:",
     "panel.settings.personal_information.button.forget_info.title": "Delete My Information",
 
-    "panel.settings.new_privacy.privacy.label.title": "Choose Your Privacy Level",
-    "panel.settings.new_privacy.privacy.label.slider_help": "Adjust slider to change your privacy level",
-    "panel.settings.new_privacy.privacy.button.set_privacy.disabled.title": "Scroll to Review",
-    "panel.settings.new_privacy.privacy.button.set_privacy.title": "Set My Privacy",
-    "panel.settings.new_privacy.privacy.button.set_privacy.disabled.hint": "",
-    "panel.settings.new_privacy.privacy.button.set_privacy.hint": "",
-    "panel.settings.new_privacy.privacy.dialog.update_privacy.title": "Your New \nPrivacy Level",
-    "panel.settings.new_privacy.privacy.dialog.update_privacy.button.back.title": "Back",
-    "panel.settings.new_privacy.privacy.dialog.label.new_privacy": "Privacy Level: ",
-    "panel.settings.new_privacy.privacy.dialog.update_privacy.label2": "This change requires us to make the following changes where applicable:",
-    "panel.settings.new_privacy.privacy.dialog.update_privacy.confirm": "Are you sure?",
-    "panel.settings.new_privacy.privacy.dialog.update_privacy.yes": "Yes",
-    "panel.settings.new_privacy.privacy.dialog.update_privacy.no": "No",
-    "panel.settings.new_privacy.privacy.label.set_your_privacy_level": "Set your privacy level",
-    "panel.settings.new_privacy.privacy.label.update_your_privacy_level": "Update your privacy level",
-    "panel.settings.new_privacy.privacy.label.set_your_privacy_level.hint": "Header 1",
-    "panel.settings.new_privacy.privacy.label.update_your_privacy_level.hint": "Header 1",
-    "panel.settings.new_privacy.privacy.label.some_details_have_changed": "Some details have changed",
-    "panel.settings.new_privacy.privacy.label.some_details_have_changed.hint": "Header 2",
-    "panel.settings.new_privacy.label.description.title": "Features and Data Collection",
-    "panel.settings.new_privacy.label.description.info": "Learn more about specific features, and use dropdown for more information about how data is being used.",
-    "panel.settings.new_privacy.button.close_all.title": "Close All",
-    "panel.settings.new_privacy.button.expand_all.title": "Expand All",
-    "panel.settings.new_privacy.label.hint": "Double tap to ",
-    "panel.settings.new_privacy.button.expand_data.title": "See Data Usage",
-    "panel.settings.new_privacy.privacy.button.set_privacy.slider.hint": "Privacy Level",
-    "panel.settings.new_privacy.privacy.button.set_privacy.slider.increase": "increased to",
-    "panel.settings.new_privacy.privacy.button.set_privacy.slider.decrease": "decreased to",
+    "panel.settings.privacy.privacy.label.title": "Choose Your Privacy Level",
+    "panel.settings.privacy.privacy.label.slider_help": "Adjust slider to change your privacy level",
+    "panel.settings.privacy.privacy.button.set_privacy.disabled.title": "Scroll to Review",
+    "panel.settings.privacy.privacy.button.set_privacy.title": "Set My Privacy",
+    "panel.settings.privacy.privacy.button.set_privacy.disabled.hint": "",
+    "panel.settings.privacy.privacy.button.set_privacy.hint": "",
+    "panel.settings.privacy.privacy.dialog.update_privacy.title": "Your New \nPrivacy Level",
+    "panel.settings.privacy.privacy.dialog.update_privacy.button.back.title": "Back",
+    "panel.settings.privacy.privacy.dialog.label.new_privacy": "Privacy Level: ",
+    "panel.settings.privacy.privacy.dialog.update_privacy.label2": "This change requires us to make the following changes where applicable:",
+    "panel.settings.privacy.privacy.dialog.update_privacy.confirm": "Are you sure?",
+    "panel.settings.privacy.privacy.dialog.update_privacy.yes": "Yes",
+    "panel.settings.privacy.privacy.dialog.update_privacy.no": "No",
+    "panel.settings.privacy.privacy.label.set_your_privacy_level": "Set your privacy level",
+    "panel.settings.privacy.privacy.label.update_your_privacy_level": "Update your privacy level",
+    "panel.settings.privacy.privacy.label.set_your_privacy_level.hint": "Header 1",
+    "panel.settings.privacy.privacy.label.update_your_privacy_level.hint": "Header 1",
+    "panel.settings.privacy.privacy.label.some_details_have_changed": "Some details have changed",
+    "panel.settings.privacy.privacy.label.some_details_have_changed.hint": "Header 2",
+    "panel.settings.privacy.label.description.title": "Features and Data Collection",
+    "panel.settings.privacy.label.description.info": "Learn more about specific features, and use dropdown for more information about how data is being used.",
+    "panel.settings.privacy.button.close_all.title": "Close All",
+    "panel.settings.privacy.button.expand_all.title": "Expand All",
+    "panel.settings.privacy.label.hint": "Double tap to ",
+    "panel.settings.privacy.button.expand_data.title": "See Data Usage",
+    "panel.settings.privacy.privacy.button.set_privacy.slider.hint": "Privacy Level",
+    "panel.settings.privacy.privacy.button.set_privacy.slider.increase": "increased to",
+    "panel.settings.privacy.privacy.button.set_privacy.slider.decrease": "decreased to",
 
     "panel.settings.video_tutorial.header.title": "Video Tutorial",
     "panel.settings.video_tutorial.video.missing.msg": "Missing video",
