+++ assets/strings.zh.json
@@ -801,30 +801,6 @@
 	"panel.events_schedule.button.details.title": "细节",
 	"panel.events_schedule.button.details.hint": "",
 
-	"panel.settings.privacy.label.title":"我的隐私",
-	"panel.settings.privacy.label.set_your_privacy_level":"设置您的隐私级别",
-	"panel.settings.privacy.label.set_your_privacy_level.hint":"标题1",
-	"panel.settings.privacy.label.update_your_privacy_level":"更新您的隱私級別",
-	"panel.settings.privacy.label.update_your_privacy_level.hint": "标题1",
-	"panel.settings.privacy.label.some_details_have_changed": "一些細節已經改變",
-	"panel.settings.privacy.label.some_details_have_changed.hint": "标题2",
-	"panel.settings.privacy.label.category.hint":"标题",
-	"panel.settings.privacy.label.description":"使用下面的滑块选择您的隐私级别。您可以随时查看共享的信息并更改设置。",
-	"panel.settings.privacy.label.bottom_description":"无论隐私级别如何，匿名数据都会被收集并用于优化性能",
-	"panel.settings.privacy.button.set_privacy.title":"设置我的隐私",
-	"panel.settings.privacy.button.set_privacy.hint":"设置我的隐私",
-	"panel.settings.privacy.button.set_privacy.disabled.title":"滚动查看",
-	"panel.settings.privacy.button.set_privacy.disabled.hint":"滚动审阅",
-	"panel.settings.privacy.button.set_privacy.slider.hint":"隐私级别",
-	"panel.settings.privacy.button.set_privacy.slider.increase":"增加至",
-	"panel.settings.privacy.button.set_privacy.slider.decrease":"减少为",
-	"panel.settings.privacy.dialog.update_privacy.title":"最新隐私设置",
-	"panel.settings.privacy.dialog.update_privacy.label1":"你的最新隐私设置",
-	"panel.settings.privacy.dialog.update_privacy.label2":"要求我们进行以下更改：",
-	"panel.settings.privacy.dialog.update_privacy.confirm":"你确定吗？",
-	"panel.settings.privacy.dialog.update_privacy.yes":"是",
-	"panel.settings.privacy.dialog.update_privacy.no":"否",
-
 	"panel.settings.manage_interests.title":"管理我的兴趣",
 	"panel.settings.manage_interests.instructions.tap":"点击",
 	"panel.settings.manage_interests.instructions.follow":"跟随您最感兴趣的标签",
@@ -1227,34 +1203,34 @@
 	"panel.settings.personal_information.label.delete_message.description2":"永远 ",
 	"panel.settings.personal_information.label.delete_message.description3":"删除:",
 	"panel.settings.personal_information.button.forget_info.title": "Delete My Information",
-	"panel.settings.new_privacy.privacy.label.title": "选择您的隐私级别",
-	"panel.settings.new_privacy.privacy.label.slider_help": "调整滑块以更改您的隐私级别",
-	"panel.settings.new_privacy.privacy.button.set_privacy.disabled.title": "滚动查看",
-	"panel.settings.new_privacy.privacy.button.set_privacy.title": "设置我的隐私",
-	"panel.settings.new_privacy.privacy.button.set_privacy.disabled.hint": "",
-	"panel.settings.new_privacy.privacy.button.set_privacy.hint": "",
-	"panel.settings.new_privacy.privacy.dialog.update_privacy.title": "你最新的 \n隐私级别",
-	"panel.settings.new_privacy.privacy.dialog.update_privacy.button.back.title": "后退",
-	"panel.settings.new_privacy.privacy.dialog.label.new_privacy": "隐私级别: ",
-	"panel.settings.new_privacy.privacy.dialog.update_privacy.label2": "此更改要求我們在適用的情況下進行以下更改：",
-	"panel.settings.new_privacy.privacy.dialog.update_privacy.confirm": "你确定吗?",
-	"panel.settings.new_privacy.privacy.dialog.update_privacy.yes": "是",
-	"panel.settings.new_privacy.privacy.dialog.update_privacy.no": "否",
-	"panel.settings.new_privacy.privacy.label.set_your_privacy_level": "设置您的隐私级别",
-	"panel.settings.new_privacy.privacy.label.update_your_privacy_level": "更新您的隐私级别",
-	"panel.settings.new_privacy.privacy.label.set_your_privacy_level.hint": "标题 1",
-	"panel.settings.new_privacy.privacy.label.update_your_privacy_level.hint": "标题 1",
-	"panel.settings.new_privacy.privacy.label.some_details_have_changed": "一些细节已经改变",
-	"panel.settings.new_privacy.privacy.label.some_details_have_changed.hint": "标题 2",
-	"panel.settings.new_privacy.label.description.title": "特性和数据收集",
-	"panel.settings.new_privacy.label.description.info": "了解有关特定功能的更多信息，并使用下拉菜单获取有关如何使用数据的更多信息.",
-	"panel.settings.new_privacy.button.close_all.title": "全部关闭",
-	"panel.settings.new_privacy.button.expand_all.title": "全部展开",
-	"panel.settings.new_privacy.label.hint": "双击以 ",
-	"panel.settings.new_privacy.button.expand_data.title": "查看数据使用",
-	"panel.settings.new_privacy.privacy.button.set_privacy.slider.hint": "隐私级别",
-	"panel.settings.new_privacy.privacy.button.set_privacy.slider.increase": "增加到",
-	"panel.settings.new_privacy.privacy.button.set_privacy.slider.decrease": "减少到",
+	"panel.settings.privacy.privacy.label.title": "选择您的隐私级别",
+	"panel.settings.privacy.privacy.label.slider_help": "调整滑块以更改您的隐私级别",
+	"panel.settings.privacy.privacy.button.set_privacy.disabled.title": "滚动查看",
+	"panel.settings.privacy.privacy.button.set_privacy.title": "设置我的隐私",
+	"panel.settings.privacy.privacy.button.set_privacy.disabled.hint": "",
+	"panel.settings.privacy.privacy.button.set_privacy.hint": "",
+	"panel.settings.privacy.privacy.dialog.update_privacy.title": "你最新的 \n隐私级别",
+	"panel.settings.privacy.privacy.dialog.update_privacy.button.back.title": "后退",
+	"panel.settings.privacy.privacy.dialog.label.new_privacy": "隐私级别: ",
+	"panel.settings.privacy.privacy.dialog.update_privacy.label2": "此更改要求我們在適用的情況下進行以下更改：",
+	"panel.settings.privacy.privacy.dialog.update_privacy.confirm": "你确定吗?",
+	"panel.settings.privacy.privacy.dialog.update_privacy.yes": "是",
+	"panel.settings.privacy.privacy.dialog.update_privacy.no": "否",
+	"panel.settings.privacy.privacy.label.set_your_privacy_level": "设置您的隐私级别",
+	"panel.settings.privacy.privacy.label.update_your_privacy_level": "更新您的隐私级别",
+	"panel.settings.privacy.privacy.label.set_your_privacy_level.hint": "标题 1",
+	"panel.settings.privacy.privacy.label.update_your_privacy_level.hint": "标题 1",
+	"panel.settings.privacy.privacy.label.some_details_have_changed": "一些细节已经改变",
+	"panel.settings.privacy.privacy.label.some_details_have_changed.hint": "标题 2",
+	"panel.settings.privacy.label.description.title": "特性和数据收集",
+	"panel.settings.privacy.label.description.info": "了解有关特定功能的更多信息，并使用下拉菜单获取有关如何使用数据的更多信息.",
+	"panel.settings.privacy.button.close_all.title": "全部关闭",
+	"panel.settings.privacy.button.expand_all.title": "全部展开",
+	"panel.settings.privacy.label.hint": "双击以 ",
+	"panel.settings.privacy.button.expand_data.title": "查看数据使用",
+	"panel.settings.privacy.privacy.button.set_privacy.slider.hint": "隐私级别",
+	"panel.settings.privacy.privacy.button.set_privacy.slider.increase": "增加到",
+	"panel.settings.privacy.privacy.button.set_privacy.slider.decrease": "减少到",
 
     "panel.settings.video_tutorial.header.title": "視頻教程",
     "panel.settings.video_tutorial.video.missing.msg": "缺少視頻",
